Fix: Extract api logic from logout button in pages component.
